Move matchers to own package to prevent import loops
